requirements.txt
pandas==2.2.3
tqdm==4.67.1
mistralai==1.7.0
PyMuPDF==1.25.5
python-docx==1.1.2
python-pptx==1.0.2
Pillow==11.2.1
google-generativeai==0.6.0
fastapi==0.115.12
uvicorn[standard]
pydantic==2.11.3
python-multipart==0.0.20
langchain-experimental==0.3.4
langchain-mistralai==0.2.10
langgraph==0.3.31
beautifulsoup4==4.13.3
numpy==1.26.4
faiss-cpu==1.10.0
openpyxl==3.1.5
langchain
camelot-py
flask==2.3.3
werkzeug==2.3.7
posthog
neo4j
spacy>=3.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
gunicorn==21.2.0